perf(interpreter): reduce DecodedInst from 48 to 40 bytes
Change bb_gas_cost from u64 to u32 in DecodedInst and all related
block_gas_costs vectors. Gas costs per basic block are bounded by
~25 instructions × ~100 cycles ≈ 2500, well within u32 range
(flush_and_get_cost already returns u32).

The 16.7% size reduction improves cache utilization for the interpreter's
instruction dispatch loop, which iterates over a contiguous DecodedInst
array. A compile-time size assertion guards against future regressions.

+++ grey/crates/javm/src/backend.rs
@@ -30,7 +30,7 @@ pub struct InterpreterProgram {
     /// Valid branch/jump landing targets.
     pub basic_block_starts: Vec<bool>,
     /// Per-gas-block costs (indexed by block start PC).
-    pub block_gas_costs: Vec<u64>,
+    pub block_gas_costs: Vec<u32>,
     /// Instruction bytecode (kept for step/trace fallback).
     pub code: Vec<u8>,
     /// Opcode bitmask.
